refactor: 트랜잭션 관리를 위해 멤버 서비스 분리
