+++ src/main/java/com/samhap/kokomen/auth/service/AuthService.java
@@ -2,9 +2,8 @@ package com.samhap.kokomen.auth.service;
 
 import com.samhap.kokomen.auth.external.KakaoOAuthClient;
 import com.samhap.kokomen.auth.external.dto.KakaoUserInfoResponse;
-import com.samhap.kokomen.auth.external.dto.MemberResponse;
-import com.samhap.kokomen.member.domain.Member;
-import com.samhap.kokomen.member.repository.MemberRepository;
+import com.samhap.kokomen.member.service.MemberService;
+import com.samhap.kokomen.member.service.dto.MemberResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,13 +12,11 @@ import org.springframework.stereotype.Service;
 public class AuthService {
 
     private final KakaoOAuthClient kakaoOAuthClient;
-    private final MemberRepository memberRepository;
+    private final MemberService memberService;
 
     public MemberResponse kakaoLogin(String code) {
         KakaoUserInfoResponse kakaoUserInfoResponse = kakaoOAuthClient.requestKakaoUserInfo(code);
-        Member member = memberRepository.findByKakaoId(kakaoUserInfoResponse.id())
-                .orElseGet(() -> memberRepository.save(new Member(kakaoUserInfoResponse.id(), kakaoUserInfoResponse.kakaoAccount().profile().nickname())));
 
-        return new MemberResponse(member);
+        return memberService.findOrCreateByKakaoId(kakaoUserInfoResponse.id(), kakaoUserInfoResponse.kakaoAccount().profile().nickname());
     }
 }
