+++ src/main/java/com/samhap/kokomen/member/service/MemberService.java
@@ -1,0 +1,23 @@
+package com.samhap.kokomen.member.service;
+
+import com.samhap.kokomen.member.domain.Member;
+import com.samhap.kokomen.member.repository.MemberRepository;
+import com.samhap.kokomen.member.service.dto.MemberResponse;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class MemberService {
+
+    private final MemberRepository memberRepository;
+
+    @Transactional
+    public MemberResponse findOrCreateByKakaoId(Long kakaoId, String nickname) {
+        Member member = memberRepository.findByKakaoId(kakaoId)
+                .orElseGet(() -> memberRepository.save(new Member(kakaoId, nickname)));
+
+        return new MemberResponse(member);
+    }
+}
